cloud_functions/transform/src/requirements.txt
functions-framework==3.*
requests==2.31.0
google-cloud-bigquery==3.5.0